fixes update interval on start
Data is fetched directly on start instead of waiting 5 seconds. However this may lead to 3-4 extra requests in the first second that may cause a 429 (too many requests) error from the server. Nevertheless it does not affect the functionality down the road. We need to fix this with something solid. Maybe use redux-saga ?
+ updates doantion address